Changed ClaimRouteCommand, UpdateCarCountCommand, ServerFacade (and iServer), and TrainCardDeck
ClaimRouteCommand sends a Route rather than an iRoute with the updated
Route model

UpdateCarCountCommand is now UpdateCarCountAndHandCommand because when
you claim a route, the server should send back a command to
successfully update the model rather than the model updating itself

ServerFacade made some updates to properly update scoreboards and added
a list of cardsUsedToClaimRoute so that those cards can be put in the
discard pile on the server side. This list will have to be generated on
the client side and sent over. It also ends the players turn if they
pick a RAINBOW card on the first draw from the face up cards, but not
the deck.

Then, I updated the TrainCardDeck model so that when the deck is empty,
it shuffles the discard pile and refills the deck with those cards